Move SimpleQueue to be a top-level class
